fix: resolved cylic dependency for snapshot package by splitting app-sdk defra files into a separate package
