+++ pkg/defrasdk/defra.go
@@ -1,4 +1,4 @@
-package defra
+package defrasdk
 
 import (
 	"bytes"
